Several changes to remove Clang warning messages related to 64 bit architectures.  Changes include:
1. Promoting signed int to ptrdiff_t
2. Promoting unsigned int to size_t
3. Adding asserts when larger sized types are cast to lower sized one
   to ensure that stored values don't exceed the range of the type.
4. unsigned int hnr replaced by HashNumber hnr in few places where missed.
5. Forced AT_64BIT to be defined or produce a compile time error.

+++ aterm/aterm/afun.c
@@ -239,17 +239,17 @@ int AT_printSymbol(AFun fun, FILE *f)
 }
 
 /*}}}  */
-/*{{{  int AT_writeAFun(AFun fun, byte_writer *writer) */
+/*{{{  size_t AT_writeAFun(AFun fun, byte_writer *writer) */
 
 /**
   * Print an afun.
   */
 
-int AT_writeAFun(AFun fun, byte_writer *writer)
+size_t AT_writeAFun(AFun fun, byte_writer *writer)
 {
   SymEntry entry = at_lookup_table[fun];
   char *id = entry->name;
-  int size = 0;
+  size_t size = 0;
 
   if (IS_QUOTED(entry->header)) {
     /* This function symbol needs quotes */
@@ -290,7 +290,7 @@ int AT_writeAFun(AFun fun, byte_writer *writer)
   } else {
     size_t len = strlen(id);
     assert(len+size <= UINT_MAX);
-    size += write_bytes(id, (unsigned int)len, writer);
+    size += write_bytes(id, len, writer);
   }
   return size;
 }
